Allow local files to be ingested
Use the provided pipe to determine which stage the files should be PUT
into, put the files into the stage, then add them to the list of files
to ingest with the Snowpipe REST APIs.

The ingest command handles any combination of already staged files and
local files in a single invocation.